vfio-bindings: Rename `bindings_v5_0_0` module
Rename `bindings_v5_0_0` module to `vfio_bindings` as we only want to
maintain one version of `vfio_bindings` simultaneously.
